Base class for S3 testing
Moves the setting up of the mock and s3 client into a base class so
that it can be shared by multiple tests.

Includes a switch, that can be flipped with the system property
testRealAWS(i.e. -DtestRealAWS=true) to run the test against real AWS,
rather than the mock. This is disabled by default, but allows us to
validate that the mock is behaving the same way as S3 would. It
requires that credentials and a default region are specified in
~/.aws.

Master Issue: #1511
